requirements.txt
pyinstaller
Flask
Flask-SQLAlchemy
mysqlclient
flask-login
sqlalchemy-datatables